+++ src/entites/player.rs
@@ -1,4 +1,4 @@
-use serde::{ Serialize, Deserialize };
+use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Player {
@@ -9,34 +9,34 @@ pub struct Player {
     name: String,
 
     #[serde(rename = "townHallLevel")]
-    town_hall_level: i64,
+    town_hall_level: i8,
 
     #[serde(rename = "expLevel")]
-    exp_level: i64,
+    exp_level: i32,
 
     #[serde(rename = "trophies")]
-    trophies: i64,
+    trophies: i32,
 
     #[serde(rename = "bestTrophies")]
-    best_trophies: i64,
+    best_trophies: i32,
 
     #[serde(rename = "warStars")]
-    war_stars: i64,
+    war_stars: i32,
 
     #[serde(rename = "attackWins")]
-    attack_wins: i64,
+    attack_wins: i32,
 
     #[serde(rename = "defenseWins")]
-    defense_wins: i64,
+    defense_wins: i32,
 
     #[serde(rename = "versusTrophies")]
-    versus_trophies: i64,
+    versus_trophies: i32,
 
     #[serde(rename = "bestVersusTrophies")]
-    best_versus_trophies: i64,
+    best_versus_trophies: i32,
 
     #[serde(rename = "versusBattleWins")]
-    versus_battle_wins: i64,
+    versus_battle_wins: i32,
 
     #[serde(rename = "role")]
     role: String,
@@ -45,10 +45,10 @@ pub struct Player {
     war_preference: String,
 
     #[serde(rename = "donations")]
-    donations: i64,
+    donations: i32,
 
     #[serde(rename = "donationsReceived")]
-    donations_received: i64,
+    donations_received: i32,
 
     #[serde(rename = "clan")]
     clan: Option<Clan>,
@@ -57,7 +57,7 @@ pub struct Player {
     achievements: Vec<Achievement>,
 
     #[serde(rename = "versusBattleWinCount")]
-    versus_battle_win_count: i64,
+    versus_battle_win_count: i32,
 
     #[serde(rename = "labels")]
     labels: Vec<Option<Label>>,
@@ -77,7 +77,7 @@ pub struct Label {
     name: String,
 
     #[serde(rename = "id")]
-    id: i64,
+    id: i32,
 
     #[serde(rename = "iconUrls")]
     icon_urls: Vec<String>,
@@ -89,10 +89,10 @@ pub struct Hero {
     name: String,
 
     #[serde(rename = "level")]
-    level: i64,
+    level: i32,
 
     #[serde(rename = "maxLevel")]
-    max_level: i64,
+    max_level: i32,
 
     #[serde(rename = "village")]
     village: Village,
@@ -103,13 +103,13 @@ pub struct Achievement {
     name: String,
 
     #[serde(rename = "stars")]
-    stars: i64,
+    stars: i32,
 
     #[serde(rename = "value")]
-    value: i64,
+    value: i32,
 
     #[serde(rename = "target")]
-    target: i64,
+    target: i32,
 
     #[serde(rename = "info")]
     info: String,
@@ -130,7 +130,7 @@ pub struct Clan {
     name: String,
 
     #[serde(rename = "clanLevel")]
-    clan_level: i64,
+    clan_level: i8,
 
     #[serde(rename = "badgeUrls")]
     badge_urls: BadgeUrls,
@@ -154,10 +154,10 @@ pub struct Spell {
     name: String,
 
     #[serde(rename = "level")]
-    level: i64,
+    level: i32,
 
     #[serde(rename = "maxLevel")]
-    max_level: i64,
+    max_level: i32,
 
     #[serde(rename = "village")]
     village: Village,
@@ -169,10 +169,10 @@ pub struct Troop {
     name: String,
 
     #[serde(rename = "level")]
-    level: i64,
+    level: i32,
 
     #[serde(rename = "maxLevel")]
-    max_level: i64,
+    max_level: i32,
 
     #[serde(rename = "village")]
     village: Village,
@@ -187,41 +187,41 @@ pub enum Village {
     Home,
 }
 
-impl Player     {
+impl Player {
     pub fn tag(&self) -> &str {
         &self.tag
     }
     pub fn name(&self) -> &str {
         &self.name
     }
-    pub fn town_hall_level(&self) -> i64 {
+    pub fn town_hall_level(&self) -> i8 {
         self.town_hall_level
     }
-    pub fn exp_level(&self) -> i64 {
+    pub fn exp_level(&self) -> i32 {
         self.exp_level
     }
-    pub fn trophies(&self) -> i64 {
+    pub fn trophies(&self) -> i32 {
         self.trophies
     }
-    pub fn best_trophies(&self) -> i64 {
+    pub fn best_trophies(&self) -> i32 {
         self.best_trophies
     }
-    pub fn war_stars(&self) -> i64 {
+    pub fn war_stars(&self) -> i32 {
         self.war_stars
     }
-    pub fn attack_wins(&self) -> i64 {
+    pub fn attack_wins(&self) -> i32 {
         self.attack_wins
     }
-    pub fn defense_wins(&self) -> i64 {
+    pub fn defense_wins(&self) -> i32 {
         self.defense_wins
     }
-    pub fn versus_trophies(&self) -> i64 {
+    pub fn versus_trophies(&self) -> i32 {
         self.versus_trophies
     }
-    pub fn best_versus_trophies(&self) -> i64 {
+    pub fn best_versus_trophies(&self) -> i32 {
         self.best_versus_trophies
     }
-    pub fn versus_battle_wins(&self) -> i64 {
+    pub fn versus_battle_wins(&self) -> i32 {
         self.versus_battle_wins
     }
     pub fn role(&self) -> &str {
@@ -230,10 +230,10 @@ impl Player     {
     pub fn war_preference(&self) -> &str {
         &self.war_preference
     }
-    pub fn donations(&self) -> i64 {
+    pub fn donations(&self) -> i32 {
         self.donations
     }
-    pub fn donations_received(&self) -> i64 {
+    pub fn donations_received(&self) -> i32 {
         self.donations_received
     }
     pub fn clan(&self) -> &Option<Clan> {
@@ -242,7 +242,7 @@ impl Player     {
     pub fn achievements(&self) -> &Vec<Achievement> {
         &self.achievements
     }
-    pub fn versus_battle_win_count(&self) -> i64 {
+    pub fn versus_battle_win_count(&self) -> i32 {
         self.versus_battle_win_count
     }
     pub fn labels(&self) -> &Vec<Option<Label>> {
@@ -266,7 +266,7 @@ impl Clan {
     pub fn name(&self) -> &str {
         &self.name
     }
-    pub fn clan_level(&self) -> i64 {
+    pub fn clan_level(&self) -> i8 {
         self.clan_level
     }
     pub fn badge_urls(&self) -> &BadgeUrls {
@@ -290,10 +290,10 @@ impl Spell {
     pub fn name(&self) -> &str {
         &self.name
     }
-    pub fn level(&self) -> i64 {
+    pub fn level(&self) -> i32 {
         self.level
     }
-    pub fn max_level(&self) -> i64 {
+    pub fn max_level(&self) -> i32 {
         self.max_level
     }
     pub fn village(&self) -> &Village {
@@ -305,13 +305,13 @@ impl Achievement {
     pub fn name(&self) -> &str {
         &self.name
     }
-    pub fn stars(&self) -> i64 {
+    pub fn stars(&self) -> i32 {
         self.stars
     }
-    pub fn value(&self) -> i64 {
+    pub fn value(&self) -> i32 {
         self.value
     }
-    pub fn target(&self) -> i64 {
+    pub fn target(&self) -> i32 {
         self.target
     }
     pub fn info(&self) -> &str {
@@ -329,10 +329,10 @@ impl Troop {
     pub fn name(&self) -> &str {
         &self.name
     }
-    pub fn level(&self) -> i64 {
+    pub fn level(&self) -> i32 {
         self.level
     }
-    pub fn max_level(&self) -> i64 {
+    pub fn max_level(&self) -> i32 {
         self.max_level
     }
     pub fn village(&self) -> &Village {
